refactor: move select core and base CSS to styles folder
