+++ packages/select/package.json
@@ -23,8 +23,8 @@
     "lit.d.ts",
     "lit.js",
     "src",
-    "!src/*-base-styles.d.ts",
-    "!src/*-base-styles.js",
+    "!src/styles/*-base-styles.d.ts",
+    "!src/styles/*-base-styles.js",
     "theme",
     "vaadin-*.d.ts",
     "vaadin-*.js",
