+++ packages/select/src/styles/vaadin-select-base-styles.d.ts
