+++ packages/select/src/styles/vaadin-select-base-styles.js
