+++ packages/select/src/styles/vaadin-select-core-styles.d.ts
