+++ packages/select/src/styles/vaadin-select-core-styles.js
