+++ packages/select/src/styles/vaadin-select-overlay-base-styles.js
