+++ packages/select/src/styles/vaadin-select-overlay-core-styles.js
