+++ packages/select/src/styles/vaadin-select-value-button-base-styles.js
