+++ packages/select/src/styles/vaadin-select-value-button-core-styles.js
