+++ packages/select/src/vaadin-select-overlay.js
@@ -9,7 +9,7 @@ import { PolylitMixin } from '@vaadin/component-base/src/polylit-mixin.js';
 import { overlayStyles } from '@vaadin/overlay/src/vaadin-overlay-styles.js';
 import { CSSInjectionMixin } from '@vaadin/vaadin-themable-mixin/css-injection-mixin.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
-import { selectOverlayStyles } from './vaadin-select-overlay-core-styles.js';
+import { selectOverlayStyles } from './styles/vaadin-select-overlay-core-styles.js';
 import { SelectOverlayMixin } from './vaadin-select-overlay-mixin.js';
 
 /**
