+++ packages/select/src/vaadin-select-value-button.js
@@ -9,7 +9,7 @@ import { defineCustomElement } from '@vaadin/component-base/src/define.js';
 import { PolylitMixin } from '@vaadin/component-base/src/polylit-mixin.js';
 import { CSSInjectionMixin } from '@vaadin/vaadin-themable-mixin/css-injection-mixin.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
-import { valueButton } from './vaadin-select-value-button-core-styles.js';
+import { valueButton } from './styles/vaadin-select-value-button-core-styles.js';
 
 /**
  * An element used internally by `<vaadin-select>`. Not intended to be used separately.
